mysite/requirements.txt
Django~=2.2.4
djangorestframework
requests